+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 1.2.0RC1
 * Move main class to includes
 * Remove Aihrus Framework loading during requirements checking
+* Require Custom Bulk/Quick Edit 1.5.0
 * RESOLVE Doesn't activate with premium plugin
 * Specify a “Text Domain” and “Domain Path”
 * Update file structure